Rename classes Titles to ActorTitles, Movies to NonAdultMovies, Local to NamesOfTitles. And made classes with a separate implemention and interface
